fix bug: filesystem changes not propagated through filterMapWritable
`unjoinMap` relies on getting a latest value when dirtied (so it can
dirty the changed keys), but we don't send a latest value from
`filterMapWritable`. In general we don't want to run map functions
eagerly for upward changes, because they might be expensive and not
actually demanded.

the fix is to add an `eager` flag on `mapInvertible` to cause it to
run the map function and send a dirty value, and use it in
`filterMapWritable` since the map function is cheap (for small maps at
least).

(we should also fix unjoinMap so it dirties all keys when dirtied
without a latest value; then it will be slower but not broken if used
on a signal that doesn't send a latest value.)