fix bug: splitMap must re-register dependency when a value changes
Normally dirtying proceeds upwards to the roots of the signal tree,
unregistering dependencies as it goes; then when the needed roots are
reconciled, dependencies are re-registered. However for `splitMap`,
dirtying might dirty signals for individual values but not the overall
map; then the `splitMap` no longer has a dependency on its underlying
map, but it doesn't get reconciled because it's not dirty. When this
happens the dependency needs to be re-registered manually.

Other nodes that pass a value upward in `dirty` (eager `map` and
`mapInvertible`) must also re-register their dependency, because some
node above might not be reconciled.